Move WpfAncestorDataBindingTest
1. Move WpfAncestorDataBindingTest
