+++ WpfAncestorDataBindingTest/WpfAncestorDataBindingTest.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.13.35931.197 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfAncestorDataBinding", "WpfAncestorDataBinding\WpfAncestorDataBinding.csproj", "{0BFB327B-2A47-43F8-8C6D-487FA792FBA4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfAncestorDataBindingTest", "WpfAncestorDataBindingTest\WpfAncestorDataBindingTest.csproj", "{21B04BA1-B8D3-4585-942A-1DE06A36BBDC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0BFB327B-2A47-43F8-8C6D-487FA792FBA4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0BFB327B-2A47-43F8-8C6D-487FA792FBA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0BFB327B-2A47-43F8-8C6D-487FA792FBA4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0BFB327B-2A47-43F8-8C6D-487FA792FBA4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{21B04BA1-B8D3-4585-942A-1DE06A36BBDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{21B04BA1-B8D3-4585-942A-1DE06A36BBDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{21B04BA1-B8D3-4585-942A-1DE06A36BBDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{21B04BA1-B8D3-4585-942A-1DE06A36BBDC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D33E5A30-611A-40EB-8294-BB47B8B25DE1}
+		SolutionGuid = {816E1601-FA12-4BD8-BF4B-23273468E8CD}
 	EndGlobalSection
 EndGlobal
