+++ WpfAncestorDataBindingTest/WpfAncestorDataBindingTest/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="WpfAncestorDataBinding.App"
+﻿<Application x:Class="WpfAncestorDataBindingTest.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:WpfAncestorDataBinding"
+             xmlns:local="clr-namespace:WpfAncestorDataBindingTest"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
