+++ WpfAncestorDataBindingTest/WpfAncestorDataBindingTest/MainWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="WpfAncestorDataBinding.MainWindow"
+﻿<Window x:Class="WpfAncestorDataBindingTest.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfAncestorDataBinding"
-        xmlns:vm="clr-namespace:WpfAncestorDataBinding.ViewModel"
+        xmlns:local="clr-namespace:WpfAncestorDataBindingTest"
+        xmlns:vm="clr-namespace:WpfAncestorDataBindingTest.ViewModel"
         mc:Ignorable="d"
         Title="MainWindow"
         Height="400"
